Colors, top bar and positioning
Have added some colors, a top bar and tried to make the positioning work, but it is not that simple. Having to nest elements in other elements just to give them a border is making it very difficult, since this breaks the way i stack information on top of each other.