Started new structure
Started the new structure but the problem is that when something needs to get information from the system, it needs to have the context that comes with being an activity. Activities can change the same xml file which is displayed in the app, BUT as different pages, not one activity changes 2 elements and another changes 1 element in the same view.