Structured Reposotory
I made a DeviceInfo class that contains an object with the relevant device info to be used in MainActivity but can also be used in other classes so I don't have to repeat the code that accesses name, os version and model. plus more can be added in the future easily if needed.