UI Works as intended
The UI now works as I want it to by connecting the edit button to the name field and connecting the approve name button to the edit name field. So now the text field becomes editable and expands as I want it to when the button is pressed and it reverts as I want it to when the user clicks somewhere else on the screen. much easier and simpler than using the ConstraintSet and ConstraintLayout